requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
dropbox==11.36.0
idna==3.4
ply==3.11
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
stone==3.3.1
urllib3==1.26.15